Remove the SQL provider from the documentation
I'm somewhat happy to report that we're finally agreeing that this
thing has fundamental flaws that make its usage unwieldy. It will
go back to development and probably return later as a mere
reference implementation, or perhaps as another implementation
layer the user can built on top of.

What the documentation used to refer to "custom mode" is, at this
point, the only recommended usage of Red Dog.