Add privacy by role type to the entity object
In order to add an specific privacy to an entity object based in the
'roles' field of the entity object, it is necessary to use the following
name convention for the privacy properties file.
entity_ROLE_{SOME_ROLE}.properties where {SOME_ROLE} is a role of the
section 10.2.4 of the RFC7483. e.g. entity_ROLE_REGISTRANT.properties.

This new way of naming the file for privacy based on the role, also
applies to vcard.properties. e.g. vcard_ROLE_REGISTRANT.properties.